get_{id,name,description}() returns const gchar*
Returning a newly-allocated string in
grl_metadata_source_get_{name,id,description}() functions was creating some
leaks.

The point is that in most of the functions, a const gchar* is more sensitive,
as usually the returned string is either used to print it, or to pass to
function that already dups it.