Use current media if metadata() is not supported
Usually, to get information from a media we are using metadata().

Nevertheless, it is possible that the source does not implement metadata(). In this case, we still have the media with the data obtained in the browse() operation (where the media was created).

Thus, if metadata() is not supported show current available information from the media (obtained in the previous browse() operation).