Add a serializing function for medias
It returns a string that represent a media.

Lot of times, having a "serialized" version of a media is very useful. For
instance, one can save this serial in a list, or send it to a non-grilo
compliant application.

Then, the media can be recovered from the serial.

Serial follows the same scheme as an URI:

  <type>://<escaped_source_name>/<escaped_id>

As it can be seen, both "id" and "source" are required in media in order to
build the serial.

<type> denotes the type of media we are dealing with. Thus, we will have:

   GrlMedia      => "grl"
   GrlMediaAudio => "grlaudio"
   GrlMediaVideo => "grlvideo"
   GrlMediaBox   => "grlbox"
   GrlMediaImage => "grlimage"

It is worth to tell that the implementation takes in account future new medias.
Thus, if in future a GrlMediaFoo type is added, then <type> will be "grlfoo"
for this new media, without needing changes in code.