Check media when resolving key dependencies
When requesting metadata() for a media, it could be possible that one of the
requested keys depends on another key which value is already set in media
(because a previous operation already set it, or because client set it
manually).

In this case, we can consider that the requested key can be satisfied, so it is
not needed to check if the dependent key can be provided by any source.