Add utility for services needing page_size + page_number
Grilo relies in offset + count to make pagination.

But there are services like Jamendo that needs the page_number and size of
page.

For these cases, a new utility function helps user to compute page_number and
offset inside the page where she needs to start looking for data. Also it
returns an "optimal" value for page size. This optimal value is the minimum
page size needed to retrieve the data in just one query to backend.