core: Add synchronous version of grl_media_source_metadata() function
Most of Grilo functions are asynchronous.

But there are some cases, when user do not want to deal with asynchronous
stuff, that having the appropriate synchronous verison would be useful.

So this patch adds this sychronous version in grl_media_source_metadata() for
free. This means that plugin author should not worry about dealing with
synchronous versions of the calls; she just implemement the asynchronous one,
as usual. Grilo will use this asynchronous version to implement a synchronous
one over it.

While running grl_media_source_metadata_sync() function, all remaining events
and idle functions are still functions, so they will be received and executed
even when waiting for the operation to be completed.