log: revamp the log system
Current log system suffers from a few drawbacks:

  * It uses Glib's default handler for all the domains, this handler should
    not be overriden by a library (imagine what can happen if every
    library depends on being able to override the default log handler),

  * You can't use the common practice to break on g_log() to trace where
    warnings come from because the decision of printing the message is
    taken in the handler itself,

  * Other gnome libraries tend to define a single glib log domain for
    the whole library.

So, instead, use a GStreamer-like logging system, with "log domains" you
have to declare and initialize.

This patch tries to keep the original author's intent by keeping:

  * the glib's debug levels and using them in the final g_logv(),

  * the same decoding of the string given to grl_log_init().

It also modifies a bit when the GRL_LOG env variable is sampled. Instead
of requiring the use to explictely call grl_log_init() to check the
GRL_LOG behing the scene, grilo now gets this variable at startup and
overrides the default verbosity level of log domains when they are
created.

grl_log_init() can still be used to set the verbosity, it has to be
called after the log domain has been initialized which, for plugins,
means after having loaded the plugin.

https://bugzilla.gnome.org/show_bug.cgi?id=627864