metadata-source: Don't try to free an uninitialised pointer
In grl_metadata_source_setup_full_resolution_mode, if the key-list is NULL
then it skips to the end of the function - which frees the list and
returns. This check happens before the list is assigned (or even, in fact,
declare), so just return and avoid running g_list_free on uninitialised
memory.

https://bugzilla.gnome.org/show_bug.cgi?id=630493