tests: Removed setUp/tearDown code in python tests
setUp/tearDown methods in the tests load and unload the plugins
so that each test is run in a clean environment.

However, the current plugin mechanism doesn't allow that, and reloading
a previously loaded/unloaded plugin produces a crash. This patch
removes setUp/tearDown code, adding the plugin load in the tests constructor
but leaving the plugins loaded during all the test suite execution. While
this is a fix which allows the tests to be run, some of them will fail.

Once that the plugin system is reworked to allow unloading/reloading, this
commit should be reverted to come back to the original behavior.