core: fix when no slow/writable keys available
When the source doesn't have slow/writable keys (empty set) the key list is
not filtered. Also, that code is repeated in each filter type.

This patch set to NULL the filtered key list, and the returned complement is
the passed key list. The code is isolated in the generic_filter() too.

Finally, added a description of the function.

Signed-off-by: Víctor Manuel Jáquez Leal <vjaquez@igalia.com>