build: fix parallel build of libs/net
Using $(top_builddir)/libs/net breaks automake's dependency between
wc-test and libgrlnet. When using recursive builds, Automake
automatically tracks the dependencies in the same directory but when
using $(top_builddir)/libs/net/libgrlnet.la it assumes the library is in
another directoy and is already build (by assuming the other directory is
before the current one in the parent's SUBDIRS).

By removing the $(top_builddir)/libs/net in wc_test_LDADD Automake now
understands that the library is in the same directory and it should be
built before wc-test and adds the dependency which fixes parallel
builds.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>