core: Add API to create relations between keys
A couple of keys are related when getting a value for one of the keys, it is
expected that the other key has also a value.

It is specially intended to be used with keys providing multivalued data: for
each single value in one key, there must be a value matching that single value.
This matched value can be NULL, meaning that it doesn't have actually a value,
but at least it is possible to create a relation between values in both keys.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>