core: Ignore errors from metadata sources in grl_media_source_metadata()
When invoking grl_source_metadata(), core can invoke several metadata sources
to get information from different keys.

It can happen that some of this sources can fail with an error. As we cannot
send several errors in a shot, the best approach is to just print a warning and
ignore the error.

From an user point of view, it means that some keys could not be solved, no
matter if due an error or because metadata cannot resolve it.

Hence, an error in grl_media_source_metadata() means that the media source is
the one who failed.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>