core: Fix filter_foo() functions
Currently, grl_metadata_source_filter_{supported,writable,slow}() functions do
not honor what the documentation states.

In order to bring more coherence here, this patch makes
grl_metadata_source_filter_{supported,writable}() to keep only the keys that
are {supported,writable}, and returning the removed keys in a new list.

In the case of grl_metadata_source_filter_slow(), it does the opposite: the
list is filtered so only the fastest keys are kept; the slower ones are
returned in a list. This is done in this way because it is the most common case
of use for this function.

This fixes https://bugzilla.gnome.org/show_bug.cgi?id=645542

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>