core: Ignore elements coming after operation has finished
In a browse/search with full-resolution support operation, each media is
invoked with resolve() to solve the keys not known by media source.

As each metadata plugin can resolve that data sooner or later than
others, the resolved medias ready to sent to the user can be in a
different order than they went sent by the media source.

In this situation, if the operation is cancelled, we wait until the last
element is ready (the one with remaining == 0), and we send this element
notifying operation has cancelled, and finishing it.

But it can happen that an out-of-order element comes after this last one
was sent.

In this case, as operation has been finished, we need to kindly discard
it.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>