core: introduce grilo operations
There are several ways to make requests on grilo at the moment. You
can either use the MediaSource API (to find medias) or the
MetadataSource API (to find metadatas) or the Multiple API (to find
medias on a set of media sources). Currently MediaSource and
MetadataSource share the same API to cancel an operation, but the
Multiple API remains different.

This patch proposes to unify all the operations and allocate their ids
at only one place. This benefits the user by making the whole API
simpler (you no longer have to remember to how to cancel your
operation, there is only one way to do it).

This other advantage of this patch is to have a centralized place
where to find all the ongoing operations (that led to fix a memory
leak earlier), in the future that could help to debug stalled
operations.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@linux.intel.com>
Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>