media-source: manage end of search/browse with splitted sources
When running in splitted modes (usually when dealing with websites),
if you get less results than what the user would like to get, we
currently end up in an infinite loop with the core of grilo keeping
asking for new items from a source which can't deliver anymore.

To manage this specific case, we just check whether the plugin has
returned a remaining results number equals to 0 as well as a null
item.

Signed-off-by: Lionel Landwerlin <lionel.g.landwerlin@linux.intel.com>