core: install plugins XML file in the same place as the plugin itself
Sticking together both library and xml file makes things easier to handle,
specially when loading plugins from a different place than standard.

So far, user and/or developer can load plugins from a different place. But this
only was affecting the library files: the xml information files were still
getting from the standard place. Adding new functions and variables to specify
where to look for the xml files adds a unneeded level of complexity.

So the simple approach is to put both files together, so the xml files can be
obtained from the same place where the library resides.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>