core: Refactor the way of loading plugins
Plugins' XML information files are loaded before the plugins themselves, so the
list of plugins and their identifiers are available to the user, who can just
load a specific subset of them.

As those XML files are just loaded once, if user unloads and reload again some
or all plugins, there is no need to re-scan all the directories, as the plugins
locations are already loaded and known.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>