core: Allow restricting the list of plugins to use
Sometimes it is useful to restrict the list of plugins to a subset of those
installed.

This patch provides an environment variable (GRL_PLUGIN_LIST) and a parameter
to Grilo-based applications to restrict the plugins to use.

Applications will only see those plugins that are in the specified list, no
matter if there are other plugins also in the same location.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>