core: Mention explicitly Grilo in the command-line options
Current command-line options affects the way Grilo uses the plugins.

But when asking for help, it only mentions 'plugins', not 'Grilo plugins'.

As these options can be added to other applications, possible with his own
pluggable system, this commit clearly specify that we meant Grilo plugins when
asking for help for these options.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>