log: Fix up GRL_DEBUG environment variable
Grilo relies on the default GLib log handler.

Due to the G_MESSAGES_DEBUG change in GLib, the default handler filters out all
the messages unless G_MESSAGES_DEBUG environment variable includes the log
domain to show ('all' for all domains).

To fix it, if GRL_DEBUG is set, Grilo log domain is added to the
G_MESSAGES_DEBUG so the messages are not filtered.

Signed-off-by: Juan A. Suarez Romero <jasuarez@igalia.com>