test-ui: Ask for all fast keys when browsing
Asking for all the fast keys in a browse/search operation is the most efficient
way if you want to show all they keys. It will take just one query to the
service to solve all those keys in a row.

For slow keys, let's use metadata() to resolve them when user clicks the item.