core: Sources always accept resolve() operation
If the source does not implement resolve(), but the user requested the
FULL_RESOLVE flag, then Grilo will use other sources to satisfy the query.

If the source does not implement resolve(), and user did not specified
FULL_RESOLVE, then the same media without changes will be sent through the
callback.

This way, when a developer wants to solve a key not supported by the source, if
it neither supports resolve() then they do not need to search for a source
supporting resolve() and the key; developer can simply invoke resolve() with
the same source, specifying FULL_RESOLVE flag so the core does the job.