core: Add "supported-media" property
This property flag for GrlSource specifies the type of media the source can
deal with.

Thus, if our application is designed to show video content only, we could
discard all sources not providing this kind of media.

The default value for this property, if not specified, is GRL_MEDIA_TYPE_ALL.