core: Skip browsed source from resolvable sources
When a key cannot be solved in a browse/search operation, we ask other
sources that implement resolve() and that can be solve that key to get
its value (only when GRL_RESOLVE_FULL is true).

But it can happen that the key is supported by the source in the
browse/search operation, only that for the result it doesn't have a
value. Thus, when looking for sources implementing resolve() for that
key, we need to skip the browsed source, as we know it doesn't have a
value for the key.