core: Add grl_deinit() function
Quitting without cleanly unregistering sources and plugins can leave unfreed
resources in the system.

This happens specially with the sources that open or create temporal files, but
as they are not cleanly unreffed the files are kept there.

So we are adding grl_deinit() function, which should be called by the
applications once they have finished to use Grilo. This function will cleanly
unload all the sources and plugins.

https://bugzilla.gnome.org/show_bug.cgi?id=710185