build: Generate AUTHORS automatically on release
Instead of generating the AUTHORS file manually, generate it automatically
based on the git commits.

The file is split in two sections: current contributors (authors that has some
line in the current code) and past contributors (those that committed code in
the past but that nowadays has been replaced by code of new authors).

Because everybody is important for this project.