grl-launch: Add tool to use Grilo from command-line
Inspired by gst-launch, this tool allows to use Grilo from command line,
invoking most of commands and features in a simple way.

Results can be saved to filename and loaded later with a spreadsheet, as
results are stored in CSV format.
grl-launch: Add tool to use Grilo from command-line

Inspired by gst-launch, this tool allows to use Grilo from command line,
invoking most of commands and features in a simple way.

Results can be saved to filename and loaded later with a spreadsheet, as
results are stored in CSV format.

https://bugzilla.gnome.org/show_bug.cgi?id=727223