core: Make sources network aware
If a source says that it requires the local network, or access to
the Internet to work properly, automatically hide it when the network
becomes unavailable, and show it when the network comes back on again.

This is implemented using GNetworkManager's connectivity property.
See https://bugzilla.gnome.org/show_bug.cgi?id=664562

Sources only need to set the "net:local" or "net:internet" tags to
be shown/hidden from applications depending on the network status.

https://bugzilla.gnome.org/show_bug.cgi?id=725148