source: Only replace source ID in notify if unset
In other cases where grl_media_set_source() is used, it only sets the
source of a media if it is the special root container media, or if the
source was previously unset.

This avoids sources being overwritten when sending change notifications
when sending changes from an aggregating source (such as the one
embedded in Totem).

https://bugzilla.gnome.org/show_bug.cgi?id=747029