build: Also bump grl-net and grl-pls to 0.3
Turns out that grilo-plugins expects grl-net and grl-pls to be the same
version as grilo itself, which is likely what a number of applications
also do.

This is easier to deal with than versions changing, and matches what
GLib does for example.

This also fixes a number of lingering references to 0.2 as the major
version number.