registry: Glob pattern matching for plugin ranks
Source names can be constructed partially with unique identifiers, making
it difficult to match directly to the static identifiers in
GRL_PLUGIN_RANKS.

Introduce glob pattern matching for GRL_PLUGIN_RANKS, which allows for
example GRL_PLUGIN_RANKS="grl-lastfm-cover*:3".

Add a short explanation and example to the documentation as well.

https://bugzilla.gnome.org/show_bug.cgi?id=761695
https://gitlab.gnome.org/GNOME/grilo/merge_requests/17