registry: Add per Source configs on keyfile
As mentioned on 396a0153c8, we could add GrlConfig data per Plugin on
the config file. This commit extends the configuration to support
GrlSource configuration.

This can be done by adding the Source ID to the Group's name in the
keyfile, after the Plugin ID. This format is suggested by
specification at "Extending the format"

    https://specifications.freedesktop.org/desktop-entry-spec/desktop-entry-spec-latest.html

e.g:
 | [grl-youtube]
 | api-key=value1
 |
 | [grl-lua-factory grl-theaudiodb-cover]
 | api-key=value2

Signed-off-by: Victor Toso <victortoso@gnome.org>