tests: really fix common test failure in GrlNet
Based on intention of commit a176b9a03d04d3f1, we should be going up
to 1s of threshold to avoid failure such as:

    ERROR:../tests/lib-net.c:114:test_net_wc_throttling_cb: assertion
    failed (received_time >= op->expected_time):
    (118905943655 >= 118905946568)

This has been tested over 500 times, with different overloads in the
machine, no failure. It still should be able to find valid regressions
on the code as the most important one would be stress test and
big-delay, the later having a 5 seconds timeout to guarantee that the
network request has not been made.

Related: https://bugzilla.gnome.org/show_bug.cgi?id=774578
Signed-off-by: Victor Toso <victortoso@gnome.org>