registry: Fix source refcount when using source-{added,removed}
Both of those signals expect a GObject to be passed, and will unref the
object before returning from the signal emission. As we didn't add a
reference to the source we passed, this callback might remove the last
reference to the source, and finalize it.

This fixes a crash on startup in Totem.

See https://github.com/flathub/org.gnome.Totem/issues/6