registry: Fix g_hash_table_iter_next() warning
As we loop through the sources as listed in the hash table, there's a
good chance that the hash table will be modified as sources are added or
removed. But as we shouldn't be modifying the hash table while it's
being iterated over, the code throws a warning, and will exit the loop.

Warning: g_hash_table_iter_next: assertion 'ri->version == ri->hash_table->version' failed

Solve this by getting a list of all the sources in advance.