registry: Fix possible crasher on startup in Totem
When announcing that a source was added, the application might choose to
say "actually, I'm not interested in this source" and unref it. As we
want to be able to carry on checking some of its properties, such as
tags, we either need to check whether the source was finalized in
between calls, or keep a list of sources to be added/removed after we've
done computation.

The latter is easier to implement, so add all sources to temporary lists
to avoid the tags disappearing from under us when emitting the signals.

See https://github.com/flathub/org.gnome.Totem/issues/6