meson: Find libxml2 manually on MSVC builds
The libxml2 build files for Visual Studio do not generate pkg-config files for
us, so this update makes Meson search for it manually using cc.find_library()
and also check we have libxml/parser.h.

Also fix the pkg-config file that is being generated in this case, since we
can't pass a manually-found library object to 'requires:'.

This will require Meson 0.50.0 or later.