grilo-test-ui: Fix tool not working in Flatpak
Since commit adc11e54537aff804c6dcd8d26ade0114632157b, it's
impossible to use grilo-test-ui in the Flatpak for another
application, as it tries to request a D-Bus name that it cannot own.
Instead of using our own app-id/D-Bus name, try to own a variant of the
real app's name.