operation-options: Fix storing numeric limits
GrlOperationOptions needs to validate user input against the maximum
and minimum values of a metadata-key of numeric type. For numeric
types, the follow should always be true:

  @min_value_registered <= @min_value_user
  @min_value_user <= @max_value_user
  @max_value_user <= @max_value_registered.

So, we do implement some sort generic CLAMP for numeric types.  Sadly,
glib does not provide as of yet a g_value_cmp() [0] so we need to
write a bit more code to make this checks in Grilo today.

The followup commit adds a unit test.

Related: https://gitlab.gnome.org/GNOME/grilo/-/issues/140

[0] https://gitlab.gnome.org/GNOME/glib/-/issues/129