operation-options: Update doc on undefined behavior
As pointed out by Carlos in the issue below, we are currently calling
va_arg (args, gint) to the user parameter, that translates to 0 which
we would happily take in Grilo as correct.

For numeric types, user can simply use G_MAXINT and friends to limit
lower/upper limits. Grilo will take care of clamping this if value if
it is outside boundaries of a specific metadata-key.

Resolves: https://gitlab.gnome.org/GNOME/grilo/-/issues/140