operation-options: bypass checks for NULL GValues
grl_registry_metadata_key_clamp() is a helper to check if a given
@value is between @min and @max. Only @min and @max should be well
defined values to be compared with. We can simply bypass when @value
is NULL (which means, @value is not changed).

This comes with a unit test that shows issue #148 but also highlights
some other bugs around the code. Those will be discussed in its own
issue trackers.

Resolves: https://gitlab.gnome.org/GNOME/grilo/-/issues/148