pls: Fix thumbnail URI for remote files in Flatpak
When browsing remote files and sandboxed, the file metadata comes from
gvfs, which runs outside the sandbox. So it might look for thumbnails
and find some which are available outside the sandbox, and give us a
path which the sandboxed app can't access.

When Flatpak'ed and remote, we need to get our own thumbnail path.